Updated itunescrawler

+++ itunescrawl.py
@@ -13,11 +13,12 @@ import json
 import codecs
 import sys
 import os
+import random
 print("here");
 def loadState():
     print("loadState")
     try:
-        state_file = open( "itunes_store_state_dump_try.txt", "rb" )
+        state_file = open( "itunes_store_state_dump_try_a.pba", "rb" )
         apps_discovered = pickle.load( state_file )
         print("bgrjbrjbiwr");
         apps_pending = pickle.load( state_file )
@@ -28,22 +29,24 @@ def loadState():
         print( "A fresh start ..." )
         return [], []
 
-print("boom");
 character_encoding = 'utf-8'
 apps_discovered, apps_pending = loadState()
 count_offset = len( apps_discovered )
 apps_categories = {}
 
 start_time = datetime.now()
-print("foo");
 
 def getPageAsSoup( url ):
     print("getPageAsSoup")
     try:
         response = urllib.request.urlopen( url )
-    except urllib.error.HTTPError as e:
-        print( "HTTPError with: ", url, e )
-        return None
+    except:
+    	return None
+    #except urllib.error.HTTPError as e:
+    #    print( "HTTPError with: ", url, e )
+    #    return None
+    #except urllib.URLError as e:
+    #	return None
     the_page = response.read()
     soup = BeautifulSoup( the_page )
 
@@ -60,7 +63,7 @@ def reportProgress():
 
 def saveState():
     print("saveState")
-    state_file = open( "itunes_store_state_dump_try.txt", "wb" )
+    state_file = open( "itunes_store_state_dump_try_a.pba", "wb" )
     pickle.dump( apps_discovered, state_file )
     pickle.dump( apps_pending, state_file )
     state_file.close()
@@ -71,17 +74,20 @@ def getApps( categoryUrl ):
     print(categoryUrl)
     previous_apps = []
     start_idx = 1
+    herecnt =0
     while( True ):
+        #if herecnt == 10:
+        #	break
         url = categoryUrl + "&page=" + str( start_idx )
         print( url )
         categoryPage = getPageAsSoup( url )
-        allAppLinks = [aDiv.get( 'href' ) for aDiv in categoryPage.findAll( 'a', href = re.compile( '^https://itunes.apple.com/us/app' ) )]
+        allAppLinks = [aDiv.get( 'href' ) for aDiv in categoryPage.findAll( 'a', href = re.compile( '^https://itunes.apple.com/in/app' ) )]
         if allAppLinks == previous_apps: break
         apps_pending.extend( [appLink for appLink in allAppLinks if appLink not in apps_pending] )
         previous_apps = allAppLinks
         start_idx += 1
+        herecnt = herecnt + 1
     saveState()
-    #sys.exit(0)
 
 def getAppDetails( appUrl ):
     print("getAppDetails")
@@ -100,11 +106,15 @@ def getAppDetails( appUrl ):
     appDetails['description'] = titleDiv.find( 'p' ).getText()
     print(appDetails['description'])
     print(appUrl)
-#    sys.exit(0)
     
     titleDiv = soup.find( 'div', {'id' : 'title'} )
     print(titleDiv)
-    appDetails['title'] = titleDiv.find( 'h1' ).getText()
+    heretitle = titleDiv.find('h1').getText()
+    for character in heretitle:
+      if character in string.punctuation:
+         heretitle = heretitle.replace(character, " ")
+         
+    appDetails['title'] = heretitle #titleDiv.find( 'h1' ).getText()
     appDetails['developer'] = titleDiv.find( 'h2' ).getText()
     
     detailsDiv = soup.find( 'div', {'id' : 'left-stack'} )
@@ -112,15 +122,23 @@ def getAppDetails( appUrl ):
 
     priceDiv = detailsDiv.find( 'div', {'class' : 'price'} )
     if priceDiv: appDetails['price'] = priceDiv.getText()
+    else:
+    	appDetails['price'] = "None"
 
     categoryDiv = detailsDiv.find( 'li', {'class' : 'genre'} )
     if categoryDiv: appDetails['category'] = categoryDiv.find( 'a' ).getText()
+    else:
+    	appDetails['category'] = "None"
 
     releaseDateDiv = detailsDiv.find( 'li', {'class' : 'release-date'} )
     if releaseDateDiv: appDetails['release_date'] = releaseDateDiv.getText()
+    else:
+    	appDetails['release_date'] = "None"
 
     languageDiv = detailsDiv.find( 'li', {'class' : 'language'} )
     if languageDiv: appDetails['language'] = languageDiv.getText().split()
+    else:
+    	appDetails['language'] = "None"
 
     contentRatingDiv = detailsDiv.find( 'div', {'class' : 'app-rating'} )
     if contentRatingDiv: appDetails['content_rating'] = contentRatingDiv.getText()
@@ -130,7 +148,9 @@ def getAppDetails( appUrl ):
 
     compatibilityDiv = detailsDiv.find( 'p' )
     if compatibilityDiv: appDetails['compatibility'] = compatibilityDiv.getText()
-
+    else:
+    	appDetails['compatability'] = "None"
+		
     customerRatingDivs = detailsDiv.findAll( 'div', {'class' : 'rating', 'role': 'img'} )
     if customerRatingDivs:
         customerRating = customerRatingDivs[-1].get( 'aria-label' ).split( ',' )
@@ -148,9 +168,8 @@ def getAppDetails( appUrl ):
             elif text.endswith( 'Agreement' ): appDetails['license'] = href
 
     apps_discovered.append( appUrl )
-#    print(appDetails)
-    
-    #print(os.getcwd())
+    if os.path.isfile(appDetails['title']):
+    	return None
     f = open(appDetails['title'], 'w')
 
     f.write("#Title:\n")
@@ -180,93 +199,67 @@ def getAppDetails( appUrl ):
     
     f.write("\n#Rating:\n")
     f.write(''.join(appDetails['rating']))
-
-
-    
-    
     f.close()
-    exit(1)
 
     return appDetails
     
 def closeFileHandlers( fileHandlers ):
-    #print("closeFileHandlers")
     for v in fileHandlers.values():
         v.close()
 
 
 if __name__ == '__main__':
-    itunesStoreUrl = 'https://itunes.apple.com/us/genre/ios/id36?mt=8'
+    itunesStoreUrl = 'https://itunes.apple.com/in/genre/ios/id36?mt=8'
     mainPage = getPageAsSoup( itunesStoreUrl )
-    #print("boom")
     allCategories = []
-#    print(mainPage);
     
     for column in ['list column first', 'list column', 'list column last']:
         columnDiv = mainPage.find( 'ul', {'class' : column} )
-        #print( mainPage.find( 'ul', {'class' : column} ));
-        allCategories.extend( aDiv.get( 'href' ) for aDiv in columnDiv.findAll( 'a', href = re.compile( '^https://itunes.apple.com/us/genre' ) ) )
-      #  print(allCategories.extend(aDiv.get('href')) for aDiv in columnDiv.findAll( 'a', href = re.compile( '^https://itunes.apple.com/us/genre' ) ))
-      #  print(columnDiv.findAll( 'a', href = re.compile( '^https://itunes.apple.com/us/genre' ) ))
-    for boo in allCategories:
-       #print(boo)
-       doo=boo.split('/')
-       print(doo[5])
-    sys.exit(0);
-    #print("\n\n")
-    count =0
-    for category, alphabet in [( x, y ) for x in allCategories for y in string.ascii_uppercase]:
+        allCategories.extend( aDiv.get( 'href' ) for aDiv in columnDiv.findAll( 'a', href = re.compile( '^https://itunes.apple.com/in/genre' ) ) )
+    cnt = 0;
+    ncnt  = 1;
+    
+    #for category in allCategories: 
+    while ncnt < len(allCategories):
+        ind = -ncnt
+        ncnt = ncnt + 1
+        category = allCategories[ind]
         print("here")
         a= category.split("/")
         print(a)
         print(a[-2])
-        if not os.path.exists(a[-2]):
-            os.makedirs(a[-2])
-        #sys.exit(1)
-        os.chdir(a[-2])
-        print(os.getcwd())
-        f=open("boo",'w')
-        f.write("jsfbwrjrbfv")
-        f.close()
-        #sys.exit(0)
-        getApps( category + '&letter=' + alphabet )
-        count=count+1
-        if count==1:
-         break
-
-    fileHandlers = {}
-    count = 100
-#    print(apps_pending)
-    while apps_pending:
-        if count == 0:
-            saveState()
-            count = 100
-        count = count - 1
-
-        app = apps_pending.pop()
-        if not app: continue
-
-        try:
-            app_data = getAppDetails( app )
-        #    print(app_data)
-        except Exception as e:
-            print( app, e )
-            exit( 1 )
-
-        if not app_data:
-            continue
-
-        if not app_data['category']: app_data['category'] = 'uncategorized'
-
-        if app_data['category'].lower() not in fileHandlers:
-            fileHandlers[app_data['category'].lower()] = codecs.open( '_'.join( ["apple_appstore", app_data['category'].lower()] ), 'ab', character_encoding, buffering = 0 )
-            apps_categories[app_data['category'].lower()] = 0
-        apps_categories[app_data['category'].lower()] = apps_categories[app_data['category'].lower()] + 1
-        fileHandler = fileHandlers[app_data['category'].lower()]
-        try:
-            fileHandler.write( json.dumps( app_data ) + "\n" )
-        except Exception as e:
-            print( e )
-
-    saveState()
-    closeFileHandlers( fileHandlers )
+        if os.path.exists(a[-2]):
+       		continue;
+        cnt = cnt + 1
+        for alphabet in string.ascii_uppercase:
+        	getApps( category + '&letter=' + alphabet )
+        fileHandlers = {}
+        count = 100
+        newcnt =0
+        tot_apps = len(apps_pending)
+        to_take = tot_apps
+        
+        while to_take > 0:
+        	to_take = to_take -1
+        	if count == 0:
+        		saveState()
+        		count = 100
+        	count = count - 1
+        	app = random.choice(apps_pending)
+        	apps_pending.remove(app)
+        	if not app: continue
+        	try:
+        		if not os.path.exists(a[-2]):
+        			os.makedirs(a[-2])
+        		os.chdir(a[-2])
+        		app_data = getAppDetails( app )
+        		os.chdir("..")
+        	except Exception as e:
+        		print( app, e )
+        		exit( 1 )
+        	if not app_data:
+        		continue
+        	saveState()
+        del apps_pending[:]
+        #os.chdir("..");
+    closeFileHandlers( fileHandlers )
